# dist/api-docs.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "GC Suite Server",
		"description": "GC Suite Server API docs"
	},
	"basePath": "/api",
	"schemes": [
		"http"
	],
	"consumes": [
		"application/json"
	],
	"produces": [
		"application/json"
	],
	"paths": {
		"/users": {
			"post": {
				"tags": [
					"Users"
				],
				"description": "This api creates user",
				"operationId": "/users",
				"produces": [
					"application/json"
				],
				"parameters": [{
					"name": "User Info",
					"in": "body",
					"description": "Create user",
					"required": true,
					"schema": {
						"$ref": "#/definitions/createUser"
					}
				}],
				"responses": {
					"200": {}
				}
			},
			"put": {
				"tags": [
					"Users"
				],
				"description": "This api updates user",
				"operationId": "/users",
				"produces": [
					"application/json"
				],
				"parameters": [{
					"name": "User Info",
					"in": "body",
					"description": "Update user",
					"required": true,
					"schema": {
						"$ref": "#/definitions/updateUser"
					}
				}],
				"responses": {
					"200": {}
				}
			},
			"get": {
				"tags": [
					"Users"
				],
				"description": "This api returns matching users.",
				"operationId": "/users",
				"produces": [
					"application/json"
				],
				"parameters": [{
						"name": "_id",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "name",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "email",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "tenantId",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "offset",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {}
				}
			}
		},
...
	},
	"definitions": {
		"createUser": {
			"properties": {
				"name": {
					"required": false,
					"type": "string"
				},
				"email": {
					"required": true,
					"type": "string"
				},
				"password": {
					"required": true,
					"type": "string"
				},
				"tenantId": {
					"required": true,
					"type": "string"
				}
			}
		},
		"updateUser": {
			"properties": {
				"_id": {
					"required": false,
					"type": "string"
				},
				"name": {
					"required": false,
					"type": "string"
				},
				"email": {
					"required": false,
					"type": "string"
				},
				"password": {
					"required": false,
					"type": "string"
				},
				"tenantId": {
					"required": true,
					"type": "string"
				}
			}
		},
...
	}
}